Add frmMain. Add tag.
* add main entrence for different components
* add tag for each point

Visual Studio is sooo smart!!
Thinking if I have to add "delete point" feature, or not now.
